+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 [Full changelog](https://github.com/mozilla/glean/compare/v65.1.0...v65.1.1)
 * General
   * Remove newly added call to set test-mode in `test_reset_glean`, instead setting test-mode only in necessary tests.
+  * Make `EventMetricType`, `ObjectMetricType`, `URLMetricType` and `Ping` `Sendable`  ([#3255](https://github.com/mozilla/glean/pull/3255))
 
 # v65.1.0 (2025-09-09)